Creating aggregate net flux comparisons for biomass_swap and US_removals. Had to clip the US_removals net flux to US border as well because it covers full tiles, including areas outside the US that aren't part of the sensitivity analysis.

+++ analyses/aggregate_results_to_10_km.py
@@ -241,7 +241,7 @@ def sign_change(std_aggreg_flux, sensit_aggreg_flux, sensit_type):
             sensit_window = sensit_src.read(1, window=window)
 
             # Defaults the sign change output raster to 0
-            dst_data = np.zeros((window.height, window.width), dtype='Float64')
+            dst_data = np.zeros((window.height, window.width), dtype='Float32')
 
             # Assigns the output value based on the signs (source, sink) of the standard and sensitivity analysis.
             # No option has both windows equaling 0 because that results in the NoData values getting assigned whatever
